requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
audioread==3.0.1
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
comm==0.1.3
confection==0.1.5
contourpy==1.0.7
cryptography==42.0.7
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.11.0
cymem==2.0.8
cytoolz==0.12.3
DAWG-Python==0.7.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.1
easyocr==1.7.1
EasyProcess==1.1
einops==0.8.0
encodec==0.1.1
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.4
flatbuffers==24.3.25
floret==0.10.5
fonttools==4.39.4
fqdn==1.5.1
fsspec==2024.6.1
gast==0.6.0
gensim==4.3.2
gitdb==4.0.10
GitPython==3.1.37
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.23.4
idna==3.4
imageio==2.34.0
iniconfig==2.0.0
intel-openmp==2021.4.0
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.18.2
jeepney==0.8.0
jellyfish==1.0.4
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.14
jsonpointer==2.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.1
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.2
keras==3.4.1
keyboard==0.13.5
kiwisolver==1.4.4
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.3
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.41.0
lxml==4.9.3
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
MiniSom==2.3.2
mistune==2.0.5
mkl==2021.4.0
ml-dtypes==0.3.2
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.8
mss==9.0.1
murmurhash==1.0.10
mysql-connector-python==8.2.0
namex==0.0.8
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
ninja==1.11.1.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.0
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.3
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.11.0
playwright==1.38.0
pluggy==1.3.0
pooch==1.8.2
preshed==3.0.9
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.25.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py4j==0.10.9.7
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.7.3
pydantic_core==2.18.4
pyee==9.0.4
PyGetWindow==0.0.9
Pygments==2.15.1
pymorphy3==2.0.1
pymorphy3-dicts-ru==2.4.417150.4580142
PyMsgBox==1.0.9
pyparsing==3.0.9
pypdfium2==4.30.0
pyperclip==1.8.2
pyphen==0.15.0
pypng==0.20220715.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
PyRect==0.2.0
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.29
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
pyspark==3.5.0
pytesseract==0.3.10
pytest==7.4.2
pytest-base-url==2.0.0
pytest-playwright==0.4.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.0.1
python-json-logger==2.0.7
python-pptx==1.0.2
python-slugify==8.0.1
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
qrcode==7.4.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.0
regex==2024.5.15
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.15.2
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.11.2
seaborn==0.13.0
Send2Trash==1.8.2
sentencepiece==0.2.0
shapely==2.0.3
shellingham==1.5.4
shiboken6==6.5.2
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.7
spacy==3.7.5
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.2
sympy==1.12
tbb==2021.13.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
text-unidecode==1.3
textacy==0.13.0
tf_keras==2.17.0
thinc==8.2.4
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.2.12
tiktoken==0.3.3
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
toolz==0.12.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.19.0
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.44.2
triton==2.3.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
ultralytics==8.1.16
uri-template==1.2.0
urllib3==1.26.13
virtualenv==20.24.5
vocos==0.1.0
wasabi==1.1.3
wcwidth==0.2.6
weasel==0.4.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==3.0.3
widgetsnbextension==4.0.9
wrapt==1.16.0
XlsxWriter==3.2.0
xlwt==1.3.0
